generic-click: building framework up
second initial commit to add some more of the skeleton code.
nothing has been tested, currently no available nodes with click
dpdk are available for testing, so much of the code may need to
change based on actual testing.

comment: this code is written using python 3.6, so at a later date
I will run 3to2 so that I can write the code with types, and make
sure the typing is correct.

Lincoln