testing: making wrappers more robust again failure
this commit is an attempt at making it easier to determine
whether or not the click modifications worked by checking the
output of stderr for certain strings.

it also removes one of the print functions and folds into a more generic
version.  the header changed for check_magi_logs to accomodate the
changes.

Lincoln